+++ include/chimaera/work_orchestrator/work_orchestrator.h
@@ -31,7 +31,7 @@ typedef ABT_key TlsKey;
 
 struct BlockedTask {
   Task *task_;
-  hipc::atomic<ssize_t> block_count_;
+  hipc::atomic<int> block_count_;
 
   BlockedTask() = default;
